requirements.txt
chardet==4.0.0
cycler==0.11.0
decorator==4.4.2
numpy==1.19.5
PyYAML==5.4.1
scipy==1.5.4
torch==1.10.2
tqdm==4.62.3
matplotlib==3.3.4